Update "template generating empty table" test to html/parsoid
The "Templates: Wiki Tables: 4. Templated tags, no content" assumes that
there's an empty tbody/tr/td structure in an empty table. This behaviour
is not consistent with the behaviour indicated in the Parsoid version of
the test "Newline before table-close generates empty table row: T208619"
(the referenced issue is marked as "Known Differences"). Hence, this
patch updates this test for consistency.

Change-Id: I0892319f809fe653dd05b83c02d0ea6e72939559