Port the sync-baseconfig.js script to PHP
The sync-baseconfig.js is outdated and not working anymore
(dependencies have been dropped from the repository). This ports said
script to PHP, with a couple of feature differences:
* the script does not read any parsoid configuration anymore - which
means the timeout is not configured. This can be re-added, but this will
require some discussion as to how.
* there was some mention of a "apiURL" command-line argument which was
not documented; it is now documented and creates files as well.
There's also been a bit of refactoring compared to a 1:1 port.

The newly generated files are included for comparison/review purposes
(and so that they are updated). There's mostly whitespace modifications,
plus whatever changed since the last update. The 'languages' part of the
file is also removed during this update, in accordance with Arlo's
comments.

The js version is deleted. The wmf.sitematrix.json file is relocated
under tools/.

Change-Id: Iee267c6093bbbdc8da12aa25f307efa71ac2b93d