Stop adding newlines in manglePreprocessorResponse
Presumably these were added since categories are conventionally found on
a new line, but it's just introducing "extraneous spacing" that editors
aren't expecting.

No change to parserTest results, so hopefully this means nothing depends on
the old behaviour. However, this function isn't currently exercised in
parserTests because it only acts on the result of preprocessor calls to
MediaWiki and Parsoid uses its own native preprocessor in parserTests.

A phpunit test is added that adds some coverage.

Bug: T277760
Change-Id: Ide710923d8a4704ed260b88aa0814f65a1416888