Support Dodo/PHP8.4 DOM implementations
PHP 8.4 lets you alias "internal" classes, so swapping in the Dodo
and PHP 8.4 implementations is "just" a matter of defining appropriate
class aliases.  (Because earlier PHP didn't allow aliases of the
internal \DOM* classes, we had to use subclasses instead, which was
clunkier.)

We add DOMCompat::isStandardsMode() and turn off most DOMCompat
workarounds when that is enabled.  We check the type of the document
given, so DOMCompat will still work if it is called from 3rd party code
with a \DOMDocument even if Parsoid is using Dodo or PHP 8.4.

Bug: T393922
Change-Id: I78b0bc548ab55deb0a5d05d6656bd76370b7e059