Introduce DOMCompat::nodeName($node)
Documents in "HTML mode" use uppercase for $node->nodeName (and
$node->tagName, which is the same).  Documents in "XML mode" (and PHP's
dom extension) use lowercase.

Create a compatibility thunk for now to ensure we're always comparing
against the lowercase version.  We cache the upper-to-lower conversion
because previous benchmarking indicating this can be a performance
bottleneck.

Add a simple bin/nodenodecheck.sh script to ensure no further
instances of $node->nodeName sneak into our code while we're working
on migrating to Dodo.

Change-Id: I7579cbd07df7650f4c7105cc9dbdc87ea294bded