PHPUtils::jsonEncode: Borrow some code from FormatJson::encode in core
Parsoid uses this function almostly exclusively in debug output.
But, it is beginning to be used in HTML output (data-mw, meta tags in <head>).
So, borrow some code that ensures that JS code won't trip up on these
encoded strings.

Separately, it is worth thinking about converting core's FormatJson class
into a library and use it wherever needed (Bug: T287432).

Change-Id: Ifc6ce69364327b254877c1f65ff0d8447331051f