Add leniency for active formatting elements in AddMediaInfo
Right now, in the parse direction, the media remains as a redlink when
the unexpected structure is encountered, which isn't great.  But what's
worse is that, when roundtripping, the media is dropped entirely when
the serializer encounters it.

Here we add a bit of leniency for an understood scenario where active
formatting elements are reopened inside the wrapper after a content
model violation closed them just outside it.

A few examples isolated from the logs are,

<span><font>hello</span>[[File:Nuvola apps date.svg|50px|right]]

and

<p>''[[File:Blue flag waving.svg|30px|left]]

Bug: T286840
Change-Id: Idc2a7b1302c79b615e8a54541fdec9b990d57710