Wrap next siblings in fixUpMisnestedTagDSR
The necessity for this can best be illustrated with an example.

<a href="1">test<span>hihi<a href="2">haha</a>hoho</span></a>

will render as,

<a href="1">test<span>hihi</span></a><a href="2">haha</a>hoho

meaning that any next siblings to the nested <a> need to be wrapped for
marking as misnested.

A test is added, extracted from the task, that shows
wrapTextChildrenOfNode falsely assigning dsr to the ":" that ends up
seralizing an extra "-" from source, on master.

Bug: T287463
Change-Id: Ic626eede654f728b4ddf0f18f7a7b20be655ca48