html2wt: Centralize wikitext escaping to one place
The wikitext escaping that ought to happen in the LinkHandler
is only when the A-node emits the link content within "[[..]]"
which introduces a node-specific escaping context.

When a link has invalid target and is going to output its
link content, let SerializerState::emitChunk handle the wikitext
escaping since it has better info about the chunk's surrounding
context.

Change-Id: I2572c597e077440472b923a4d123681dde1af188