Only set up DOM aliases once
For WMF CI, we use a hack where we override the version of the library
loaded from vendor/wikimedia/parsoid with the code loaded from the
extension directory check out.  That ensures CI runs against the
latest Parsoid code, not just the "released" parsoid code in
mediawiki-vendor.

We used a hack in RegistrationHookHandler to ensure that the DOM
aliases are set up if we're running the code from the extension
directory, even if the version in vendor/wikimedia/parsoid doesn't
know about the DOM aliases yet.  But this breaks when we actually
release a version to vendor/wikimedia/parsoid that sets up the
DOM aliases; now we try to do it twice, and fail.

This is compounded by the fact that previously $wgParsoidUseDodo
*looked* like a global, but it wasn't declared with `global` so it
wasn't actually a global. So any test that looked to see if
$wgParsoidUseDodo was set would fail.

So: make $wgParsoidUseDodo *actually* a global, and tweak the
code in RegistrationHookHandler to only set up the DOM aliases
if they're not already set up.  There are still some gotchas
here since we're always going to execute
`vendor/wikimedia/parsoid/DomImpl.php`, never the one loaded
from the extension directory, but with a little bit of care
this can probably be finessed.  The $wgParsoidDomImplVersion
global has been set to try to make this a little easier.

Change-Id: I2420b5d99d0b6c57f4799cfb38e7b58bb2da9822