Use Parsoid's version of idle-dom and dodo when testing in integrated mode
One big bug fix where: we had an extra 'vendor/' prefix when we were
collecting core's versions of idle-dom and dodo to add to phan's
`exclude_file_list`.

But also clean up some of the documentation, now that (shortly) core
will have it's own copies of dodo and idle-dom, and now that we've
released a version of dodo without the problematic
`vendor/wikimedia/dodo/.phan` files (T287419).

Change-Id: Ibb3d0df4a348bdc483cce5945d30574a9be1d97f