Stop stripping trailing <nowiki/>s
Trailing <nowiki/>s are meaningful when used inside templates and
parser functions, and the regexp here does not cover all cases
correctly (see examples on the task).

The "interim fix" was originally added in 2015 in 06baaf4ff65a
as a workaround for T58381, which has been fixed properly in 2017.
(Mostly, there's a follow-up: Idcea69a0.)

Bug: T264027
Change-Id: Ie2c45a014acb515e8cd348126fef8e4e085f412e