Reduce array_reduce() usage
Progressively appending to a string or array in array_reduce() has
O(N^2) time due to the need to copy the string or array each time it is
modified.

The remaining usages of array_reduce() will be slower than a loop due to
the constant factor overhead of the closure calls.

I left the one in WikitextSerializer because the loop body is relatively
large, so the function call overhead is not so important, and because
the loop body would probably be a separate function anyway, and because
the code is so dense that it would take me another half an hour to
figure out what it is trying to do.

array_reduce() emits an alluring siren song which can only be heard by
those initiated into the dark secrets of functional programming. Please
restrain your shipmates lest they be dashed on the rocks of O(N^2)
performance.

Change-Id: I295dc63beef86de0b2f0562e0e8b5441e8305ff2