Simplify skipOnAny usage in token handlers
'onAny' is a generic / fallback handler that is meant to be invoked
(as a retry of sorts) if the more specialized handlers don't transform
the token.

So, if the token has been modified, the 'skipOnAny' signal is useless.
The ListHandler's use was along this lines.

I also simplified the PreHandler's conditional skipOnAny use to be
unconditional since the modified check takes care of skipOnAny already.

ParagraphWrapper was setting skipOnAny signal for return values from
the onAny handler! So, I got rid of those.

So, this effectively leaves us with 3 handlers that set skipOnAny.
* ParagraphWrapper uses it on EOFTk.
* PreHandler uses it on NLTk and EOFTk.

Change-Id: Ic412f701ad132a85469ff144ee9b028a4936d1e4