Don't defeat the empty-argument expansion cache
Share a PPFrame object between replaceVariables() requests so that
template expansions for the same template can be shared, if the
arguments are empty.

Parsoid can do a better job of this since it has the source text and can
hash it. But maybe it's nice to share these objects anyway. (Parsoid is
currently not caching anything.)

Change-Id: I85ab4f1133b7a1afde3c8bafe954959bcc05b7b1