Reimplement TokenHandler profiling and tracing using a proxy
Simplify TokenHandler::process by moving profiling and tracing out to a
proxy object. This is not quite as neat as I hoped it would be, but I
think it is an improvement.

The overhead of --trace=time is reduced from 50% to 30%, mostly due to
the PHPUtils::jsonEncode() call which I deferred.

Change-Id: I9fe619e24e61d8a77f6beb37d999ecc5e75ae91b