ParagraphWrapper: fix slow array_shift() loop
The array_shift() starts to dominate once there are a few thousand
consecutive newlines. I tried a couple of different approaches to
this, notably SplQueue, but it wasn't as fast for the normal case.
Eventually I realised that it's a private method and we don't really
need to have queue-like consumption since resetBuffers() is called after
the loop. We can just have an offset.

This version is about 0.2% slower in the normal case but 23% faster for
10k newlines.

Change-Id: I974b10ab71ad59e1a4d1c6f7f9ca43a4519032c3