Verify that strtolower() works for all byte values
I confirmed that strtolower() will indeed mangle non-ASCII UTF-8
characters if the locale is set to something with a legacy 8-bit
character set. I also confirmed that the previous exception is reachable
by setting the locale to Turkish.

Using strtr() or preg_replace_all() is slow and inconvenient, and leaves
the bug unfixed in the many places in MW core where strtolower() is
called. The correct place to fix this is at the platform level.

So, update the exception message and expand the checks to all non-ASCII
characters.

Bug: T291234
Change-Id: I7b40a0344c2f0b722749a477e7a7c2fdef22a30e