Add annotation tags to SiteConfig.php
This patch is a pre-patch for I5da512d997. The goal is to provide an
entry point in SiteConfig to get a list of annotations, so that we can
access it in VisualEditor and prepare VisualEditor handling of
annotation meta tags before we merge the handling of annotations in
Parsoid.

Bug: T261181
Change-Id: I188820e7a777edbc21245595f1219413dd9e58c1