Deduplicate file info requests
This also passes more than one file at a time to DataAccess::getFileInfo
which has the benefit of calling repoGroup->findFiles all at once in the
integrated case.

Also note that this fixes some bugs in getFileInfo implementations when
called with more than file in the array.  For example, in the integrated
case, if the same filename was requested with different dimensions, the
former would have been overwritten.

Bug: T251624
Change-Id: Iff145d9be691c518ec07227b479322c2c0e1bf9f