Don't try to deep clone DOM nodes
serialize(DOMNode) throws an exception on PHP 8.1. The only thing trying
to do this from parserTests was encapsulateTemplates(), which cloned
$dp1 which has $dp1->tmp->tplRanges->startElem. But $dp1 is only a
local variable, and only $dp1->dsr has a lifetime longer than the
function call, and the only assignments were to $dp1->dsr->start and
$dp1->dsr->end.

So, instead of deep cloning of $dp1, do a shallow clone of $dp1->dsr. It
is a SourceRange object which can be safely shallow cloned.

Change-Id: I1022b7a4b9485d51c65ef26d2f3ca668534eb539