Add ContentMetadataCollector interface
This interface is factored out from core's ParserOutput, and provides a
way to collect metadata during the parse, for example, the set of
images, links, and categories for the page.  A follow up patch will
instantiate this interface and make it accessible internally and to
the extension API.  It will be stubbed out when running in standalone
mode, although the metadata collected will be used to implement certain
parser tests modes which we don't currently handle.

Additional methods will be added to this interface in the future,
once a number of additional refactorings and improvements are made
to ParserOutput (primarily to break circular dependencies).

Bug: T287216
Change-Id: Ia02c6774c87b13d1ae5a8ed1e55cdd8c88c19b9e