Lazy-initialise the DataParsoid->tmp property
Add an accessor for $dp->tmp which always returns a stdClass. It's still
OK to use isset() or empty() on a tmp sub-property, so you can check for
a boolean without ever creating an object.

I'm also centralising construction of tmp so that it can easily be
converted to a special class.

Probably ~0.1% faster, judging by instruction count.

Change-Id: I23cdddc52f77b75350e423180bfb980856aa22ef