Add RemexPipeline
Replace the tuple returned by Env::fetchDocumentDispatcher() with an
object that holds all the things related to the RemexHtml pipeline. This
gives me a place to put the increasingly complex things I'm trying to
do with this pipeline, and allows public access to previously hidden
pipeline components. It allows us to move some code that is very
specific to Wt2Html\TreeBuilder out of the Env class.

I rewrote the comments about Dispatcher and the need for a Tokenizer to
reflect my view on the subject. I don't think it is temporary or fixable
in RemexHtml. See RemexPipeline::__construct().

Change-Id: I9f92070e273ef188fa372d25309979ea98635465