Fix global state of annotation id and DOMPostProcessor pass order
The issue in T295236 showed that the annotation were not necessarily
unique on the page, which led to unexpected processing during the
WrapAnnotation pass. Once this was solved, it triggered the same issue
as T295233, which had to do with the order of WrapAnnotation and
UnpackDOMFragments passes (as established by experimenting with playing
around with the order/activation of these passes). This implied putting
the annotation ID addition back into its own pass; to avoid a
performance hit due to the additional traversal on pages that do not
contain an annotation, we also introduce a conditional execution of that
pass on annotations (as well as on WrapAnnotation since it is
available.)

The specific cases pointed by T295233 and T295236 pass without error or
warnings on local parse.php; restarting rt-testing would be necessary to
check that the fix is catching all the instances.

This patch enables the WrapAnnotations Pass on template content as well
since we want annotation markers for annotation tags from all sources,
not just the top-level page.

Bug: T295233
Bug: T295236
Change-Id: Idcbf4935ed22c29d6958f19e00ec4515c688c18c