Fix DSR computation when end tag is pulled out of <p> range
The computation of the DSR range is wrong in case there's an inline tag
(such as <b> or <i>) as last child of a paragraph, when pulling a <meta>
tag back out of a paragraph. Instead of looking at the length of the
text of the last node of the paragraph, we now look at the look at the
length of the textcontent of the full paragraph when adjusting for new
lines, which fixes the issue.

Bug: T295330
Change-Id: I8aa8fef58c96472cce9429652125c065ca8f31c1