Implement basic variant link resolution
The PHP parser, after its initial link target resolution in
LinkHolderArray:: replaceInternal(), does a second pass for links whose
target was not found by attempting to resolve their language variants
(if applicable) and, in case of success, setting the href and title of
the generated link to point to the variant instead of outputting a
redlink. This patch is a rough draft of what this could look like in
Parsoid.

NB: the PHP parser performs language conversion for each batch of titles
in a single pass. It can do so safely, since the legacy language
converter does not need to know the source language variant of its input
and can therefore safely operate on a concatenated string of titles that
may contain titles of several different variants. The Parsoid language
converter is more sensitive to such inputs, so this patch currently uses
a naive implementation where each title's variants are resolved
individually.

A potential solution for batching language conversion on the Parsoid
side could involve grouping links by the language variant of their
target before processing. However, this still requires determining the
language variant of the given link. This could potentially be done by
relying on data-mw-variant-lang attributes generated by
ConversionTraverser in the LangConverter pass, but this transform might
not happen at all for various reasons, e.g. if the page is marked with
__NOCONTENTCONVERT__ or language conversion was not requested, so we
cannot rely on this data having been generated for us. That seems to
leave us with the only option of heuristically detecting source language
variant for each title, which might end up nullifying the performance
gains that batching would give us.

Bug: T258856
Bug: T53587
Change-Id: Idf7bb9769d856197c8b283d6ab30726a44892516