WikiLinkHandler: use the value offset for the fragment TSR
The key is always "mw:maybeContent" because that's the only kind of KV
that's added to contentKVs by buildLinkAttrs(). These KVs are always
generated with zero key source length. The key source length was used
as the fragment length, which ComputeDSR apparently mostly managed to
ignore, but after removal of end meta tags, the issue becomes evident.

So, use the KV value offset for the subpipeline fragment source offset.

Also, fix a typo in the option name "inPHPBlock".

Change-Id: I2eb28324d1ede9abeeee3b7d5c51084fa5d8514b