Remove scrubWikitext as an option
We're making our lives harder continuing to maintain this as an option,
just always enable it.

However, a bunch of tests need to be adapted to the new reality that
these normalizations are part the defaults and less nowiki escaping will
be visible, and harder to cover.

The hack from 6ed21f3 is maintained, though it's suspect.  The idea
seemed to be to normalize the old html to avoid unhelpful semantic
differences when comparing with the new html, that was normalized
similarly during the roundtrip.  However, it was applied
indiscriminately and wasn't well documented.  It should be tested
whether the changes here preserve the desired effect.

Bug: T297298
Change-Id: I5de9e42ab0dd36a751c19e06dc42403bd7c7ada7