Consider annotation tags as not rendering-transparent
This change solves two issues:
* annotation tags getting hoisted out of heading tags in scrubWikitext
mode
* white space getting removed between translate unit markers and text in
case of inline translate markup starting just after markup that has
optional " " separator (headings, lists, table cells...).

While it may not be obvious that considering annotation tags as not
rendering-transparent is the right call, it "protects" the inline
annotated content from normalizations that may interfere with their
existing behaviour - inline/block semantics and translation markers for
instance.

As a side effect, it also solves a bunch of triple \n in selser when
deleting paragraphs around annotation markers - hence the removal of a
bunch of knownFailures. These failing tests were in the category of
"things we can live with" but the behaviour is more correct with this
patch.

Change-Id: I21834b4b125f3bd96bb50b6e3244b1f8435fe58a