Only accept </> for tvar annotation when the page has annotations
Follow up to 1be0459. This is a workaround to avoid </tvar> replacement
of </> on wikis that do not have <translate> annotations installed: the
reasoning is that it would be fairly unlikely to have both a stray
<translate> tag and an unrelated </> in the same page. Additionally,
if no annotation has been detected on the page so far, </> cannot be a
shortcut for </tvar> (because otherwise <tvar> would have been detected
as an annotation), so this check is also useful on wikis where Translate
is installed, but where </> happens on a page marked as translatable.

Change-Id: Ia7f1fa741a6f6c227e3ae3c5afa57bb457c069b0