Don't generate nested annotated ranges in HTML output
Removes the nested annotations from the HTML output, with
the rationale that annotations only apply once to a node
(which may not be entirely true due to, for instance,
attributes on annotations.) Additionally, the only supported
annotations so far are the ones from the Translate extensions,
which do not support nesting.

It can be debated whether this check should be done in wt2html,
in html2wt, or both. This patch (in wt2html) solves the current
issue (T296169), and can be justified by the "DOM tree annotation"
semantics, but the Translate use case would probably benefit from
a html2wt direction in order to make sure we're not corrupting
translation units. (But, it shouldn't happen as long as we're
editing with VE, because VE has checks for nested <translate> tags.)

Bug: T296169
Change-Id: Ie897b0d248f99b43e515605f15ad662735237731