Fix "undefined index" on annotation nesting removal
Follow-up to 9a4f8ea after the rt-testing started complaining about
undefined index.

In (corner-case, not well-nested) structures like [[<tvar
name=v>link|link desc]]</tvar>, the </tvar> is seen without the <tvar>
annotation being open (because it's open in an attribute and hence not a
full annotation range). This patch fixes the warning that gets emitted
in that case.

T296485 has been filed as a follow-up for corner case situations that
occurred to me while looking at this fix.

Change-Id: I345aeb5faffafba1bfce7bb4268302bfa7b206dc