Treat translation unit marker comments as not-SOL transparent
Removing the SOL-transparent character of translation unit markers
prevents <pre> tags to be added to paragraphs that start with an inline
translate annotation.

This is not pretty, because it adds some more Translate-specific
code to TokenUtils, but I believe it does address the largest part of
rt-testing issues we currently see with Translate, and it's a low-cost
patch - which can help until we find a better way to handle this issue.

Bug: T296107
Change-Id: I39f41141b7768486a407b299450eeff2528dac6e