Fix parsing of interlanguage links with spaces
The major fix in this patch is WikiLinkHandler::getWikiLinkTargetInfo,
line 100, which allows to detect that we indeed have an interlanguage
link there.
The rest of the adjustments ensure that we get the correct link (without
spaces around it) and the correct link text (with spaces around it). I
must admit that my confidence that this patch does the right thing
mostly stems from the fact that it doesn't break other tests in this
area (while I broke them A Lot while I was trying to adjust everything),
and not from a deep understanding of why/how every part is necessary.
But, I tried to keep things as local as I needed them to hopefully avoid
side-effects.

Bug: T156308
Change-Id: Ic4f96eccc6623faf94dd81bcfac1f811b5e6475c