Simplify DOM class aliases.
Remove $wgParsoidUseDodo and instead define the DOM class aliases with
simple PSR-4 loaded class files.

It's still possible to swap the DOM implementation for benchmarking by
patching the class files.

With preload it's not possible to use require_once to possibly load an
alias, since the aliases are preloaded but the file is not.

Bug: T240775
Change-Id: I99e1768c4dcbe84d73d13ddd34a181d68e9fcb5e