Use MediaWiki's object factory when instantiating extensions
This ensures that extension registration can declare the use of
MediaWiki services, and have those services properly plumbed into
the extension implementation.

A companion patch will expand the extension schema in core to ensure
we can pass the 'services' property in our object factory specs.

(While we're at it, we're moving to the new namespace for
ObjectFactory, which was introduced in ObjectFactory v3.0.1.)

Change-Id: I27f6e33317898833c61d2697863d665120da27cb