Simplify bailTokens
This is less performant than what it replaces, but it is cleaner,
easier to understand, and possibly only encountered in edge cases.

Considering we're already doing this for the wikilink in bracket
scenario from I8c20093fcf41f34cd7007ac018a0f9c02b836466, which seems
like probably the most common case, the simplification might be
worthwhile.

Change-Id: Ieec52c67ae66b822a2c4779757b0d1a00a7d3ab6