Support lang= option for SVGs
The task at T208620 is about this being set based on the page language
but the explicit option didn't work either, which is what this patch
attempts to resolve.  A follow up will be for the page language.

FIXME: Do querystring params belong in the resource attribute?
This question came up in Ia94138383ebdbfc2feef75fdf651b969085a72b1
but was never resolved.

Bug: T208620
Change-Id: Ic93b1242a37ba41616937ae708ec52e05a8e5b2f