Set pipeline opts for attribute expander used in TemplateHandler
This resolves the roundtrip semantic differences that were noted for
Ieec52c67ae66b822a2c4779757b0d1a00a7d3ab6.

If the right pipeline opts aren't set and we expand a templated template
name then back out of the template token with
TemplateHandler::convertAttribsToString(), we may have introduced a
template wrapped nested template.  Without the correct pipeline opts,
the expanded name could contain transclusion meta marker.  And then,
even though the pipeline in convertAttribsToString sets the right
options, because the meta markers already exist in the template's
attributes, they'll be preserved and returned.

Why is that a problem?  Our template wrapping algorithm handles nested
templates quite nicely, you say.  Well, there's an assumption that any
template wrappers it sees *come from the top level* and the nesting is
the product of dom building (or deliberate enveloping in the case of
fostering).  This is an issue when combining the CompoundTemplateInfos
in the DOMRangeBuilder after sorting out all the overlapping ranges.
The template info for the nested ranges that don't come from the top
level are already covered by the parent template info and duplicate
entries get added to the overlapping template's "parts".

The example from metawiki is,

{{{name|{{{1|[[{{#ifexist:{{label|{{{q|}}}|en}}
  |{{label|{{{q|}}}|en}}
  |d:{{{q|}}}
}}|{{label|{{{q|}}}|en}}]]}}}}}}

In trying to isolate a more compact test case, I came up with,

{{{1|[[
 {{!}}]]}}}

Unfortunately, this failed in the same way but from a different stage in
the pipeline, during the TokenStreamPatcher.  There, we do some more
retokenizing but it's even more insidious.  The TemplateHandler
tokenizes template content and moves it along the pipeline until the end
of stage 2, with the appropriate pipeline options set.
TokenStreamPatcher is in stage 3 and so, even if we did pass the
appropriate pipeline options, they wouldn't be set correctly for tokens
in between template meta markers at that stage.

This patch introduces a number of FIXMEs where follows up and further
analysis is needed.

Change-Id: I68fb473b80e76e9e30bf5233dc111c25eed7dbd5