Don't add synthetic magic words to the output
These are listed as affecting page content but they don't need to exist
in the output to do that,
https://www.mediawiki.org/wiki/Help:Magic_words#Technical_metadata

Displaytitle wasn't implemented during the port so removing this isn't
going to change anything as it stands.  When we do get around to it, it
seems like it would set on the environment rather than putting it in the
page.

Similarly, defaultsort is set on the ParserOutput and affects how the
page is sorted on category pages it belongs to.  It doesn't need to be
in the HTML output of the page.

Follow-up to Ib5c20a6cb6877f5c26f4615ddae36520b2cd6976

Change-Id: I3438d758ce7b26fa4f1e745c03ebf56705043298