Tweak parameters for API requests to add more metadata
This allows us to test a few more methods of ContentMetadataCollector
by propagating the information from the API -- in particular,
externallinks.  In addition, when updating the API requests it turns
out that the "citation needed" template {{cn}} no longer includes a
category, so I've added an explicit category to the requested wikitext
to ensure we're still getting that metadata back in the test requests.

Change-Id: Ic0b03c081a6be020e3ca3c11c36d71de44ee129a