Only define our services if they haven't already been moved to core
This is a bit of future-proofing for an eventual move of most of the
extension code in the MWParsoid namespace to core.  It is anticipated
that the code will live in MediaWiki\Parser\Parsoid once it is in core.

By rolling this out onto the production train early we ensure that
once the code *is* moved to core we can still safely rollback to a
previous Parsoid revision, since even "old" versions of Parsoid
will know better than to (re)define Parsoid services if core's
already doing it.

Bug: T302118
Change-Id: Ice59bcc47d7d2054ce3ce18593834925a1173762