Add mw-file-description class on links to the file description page
The legacy parser calls this .image, but that's not very appropriate
should this ever be added to other media types and comes with a lot of
baggage in styling that targets it that might not be appropriate for
Parsoid output.

An alternative might be to add a class when the link *isn't* to the
file description page but that seems less clear.

This is desirable so that MultimediaViewer knows when to open.  A
previous attempt to determine this implicitly by comparing the link to
the resource failed because of inconsistencies in how the urls are
generated when querystring parameters are specified and urls are pretty
printed.  There's also a popular gadget that modifies the urls that
would need accounting for,
https://www.mediawiki.org/wiki/Snippets/Direct_imagelinks_to_Commons

Bug: T292657
Change-Id: Iccee2dcbc7b06d80bcb4e026eedc11042585550b