Do not break heading detection on annotations
This patch aims at handling the <translate>==test==</translate> syntax
to generate an annotated heading.
It required three modifications:
* allow annotation tags as SOL element in the grammar
* allow closing annotations as "things that can happen at the end of a
heading" in TokenizerUtils.php
* allow annotations in the heading detection in the grammar.

Additionally, for symmetry reasons, include limits
(noinclude/includeonly/onlyinclude) are not preventing headers to be
generated anymore - which aligns with the legacy parser's behaviour.

The unrelated tests modifications are due to the change of ID for
"about" in translate, which come from the additional calls to
maybeAnnotationOrExtensionTag - which I don't believe hurt anything, but
to be discussed (is there a cache impact?) - we may need to refine
things there.

The related test failures come from the fact that html2wt will NOT
generate <translate>==test==</translate> (but will instead add
newlines), which also breaks the selser oracle.

Bug: T296412
Change-Id: I47e65e9d476794814fd3623340880b7041bc1be0