Add logging of nested annotation tag removals
We want to have an idea of how common it is for extended annotation
ranges to nest another annotation range - adding some logging for that
purpose. We actually want to add logging rather than metrics to a/ be
able to distinguish between single parsing of multiple pages and
multiple parsings of the same page and b/ have the information of the
page where this nesting happens to have a better idea of the generating
wikitext.

Change-Id: Ie7a74916c5275c28bf250642f3339de517e48a93