PreHandler: Get rid of recursion in onAny handler
As it was written, the code was sketchily dropping the return value
of the recusive onAny call. It turns out that didn't matter because
the string token would just be buffered and an empty result was being
returned. But, that pattern was potentially fragile. This patch
eliminates the recursion as per Arlo's recommendation during review
of a a different version of this patch.

Change-Id: I7d63cd0951719abcd5e8668a714a36566bdfaef9