Remove seemingly unnecessary template target resolution
Perhaps an earlier refactor made this redundant but we just finished
resolving the template target before entering processSpecialMagicWord.
Indeed, that's how we know it's a special magic word.

The state that gets updated in resolveTemplateTarget should already be
set at this point.

Using the same arguments to TemplateExpandResult seems consistent and
appropriate.

Change-Id: I573fc6b7376f40ef8c16492b82c2f795a50e1e0d