Making Translate extended ranges uneditable - fixing bugs
Follow up on Ic9a17049b4e2ddda1fca92c5d7446e8118bfa8c9 after RT-testing
mishaps.
1/ fix the case of "ranges being edited after they have been removed",
which was the source of the crashes by marking these nodes as misnested,
and not editing ranges in that case
2/ add DSR to the wrapper to fix the newlines
3/ remove the extended annotation wrapper during DOM canonicalization so
that it doesn't interfere with the new lines handling.

With the extended annotation wrappers removed, selser[5] test failures
are all fixed.

Bug: T305066
Change-Id: I9567c57d04c0ae9acaf0a59d3bc293783ec71017