Adjust whiteline handling around table for extended annotation ranges
When an annotation range gets extended around a whole table, extraneous
lines get added in selser when the tags are put back in their original
place instead of being serialized where they appear in the HTML. This is
particularly visible when we forbid the edition of extended annotation
ranges, because selser does kick in systematically in that case. This
patch aims at fixing this situation.

Similar things can/should probably be done around other blocks, but I'd
like a first check on this one before I get lost in a corner case rabbit
hole.

This patch is technically decorrelated from the main patch of this
branch (following), and can be reviewed independently, but its main use
case is to fix selser issues in the following patch.

Change-Id: I33b98f92be806bf24013b94729dae993af8277aa