Don't rely on the link element to roundtrip the |page= option
It would be fine if we were always linking to the file description but
we aren't.  Two test cases are added here.  The first demonstrates that
with a |link= media option if $restBaseMigrationHack were false, we'd
lose the page.  The second shows that we currently don't roundtrip the
right page if |link= media option is present with its own page
querystring parameter.

If we want to support page in the html, we'd need a data-page attribute
on the media element, similar to how that lang attribute works.

This reverts some of what was done in 6ac70aa means that we can close
out T259931 as invalid.

Bug: T259931
Change-Id: Ibca71eb4310307a193096b881081feaf6b5c13f5