Update sync help instructions to guard against a naive oversight
I got bitten by this in the latest sync to Cite because after my
last sync to Cite ( I165ee24e1b78bdf181fa45430fdec1549310c359 ),
I didn't run the fetch script which meant that when I ran the sync
script, the checkout was from the same sync point as the previous
sync run. So, I ran into a bunch of merge conflicts which confused me
till I realized what was going on.

Change-Id: I1ac8498c028206064a85124a238cdfa402c484fd