Remove service-runner as a dependency
The use in ParsoidConfig was added in 60eaa76 for lib/index.js, which
doesn't exist anymore.

Launching a server with serviceWrapper.js doesn't make sense anymore,
now that the js http server has been removed.

Reducing the dependencies avoids a lot of pain in trying to keep
packages up to date with security fixes.

Change-Id: I6cfbc32b73451380a344671a81e7b7a3195aebca