Fix using sort order from data-mw when serializing template params
The bug was introduced in 6658cad while porting.

The output of the test prior to this patch, on php 7.4 is,

{{Meh|f1=foo|f10=foo|f16=foo|f15=foo|f14=foo|f13=foo|f12=foo|f11=foo
|f9=foo|f2=foo|f8=foo|f7=foo|f6=foo|f5=foo|f4=foo|f3=foo|f17=foo}}

The docs for usort say, "If two members compare as equal, they retain
their original order. Prior to PHP 8.0.0, their relative order in the
sorted array was undefined."

Without the fix, the comparator returns 0 (equal) for all members.  The
bug was only apparent on php versions prior to 8 and when the number of
elements in the array was greater than 16, which I guess triggers some
change in the backing store and where the undefined behaviour came from.

Bug: T304730
Change-Id: I72033e4a7cd931fc2bc5628460b790ced14aede0